// frontend/src/app/shared/api/model/get-me-restaurant200-response.ts
/**
 * Food Driver API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { GetMeRestaurant200ResponseMenuInner } from './get-me-restaurant200-response-menu-inner';
import { GetMeRestaurant200ResponseCategory } from './get-me-restaurant200-response-category';
import { GetMe200ResponseStreet } from './get-me200-response-street';
import { GetMeRestaurant200ResponseIdImage } from './get-me-restaurant200-response-id-image';
import { GetMe200ResponsePostalCode } from './get-me200-response-postal-code';
import { GetMeRestaurant200ResponseDescription } from './get-me-restaurant200-response-description';
import { GetMeRestaurant200ResponseImage } from './get-me-restaurant200-response-image';


export interface GetMeRestaurant200Response { 
  [key: string]: any | any;


    /**
     * Unikalny identyfikator
     */
    idRestaurant: number;
    name: GetMe200ResponseStreet;
    description: GetMeRestaurant200ResponseDescription;
    category: GetMeRestaurant200ResponseCategory;
    idImage?: GetMeRestaurant200ResponseIdImage;
    image: GetMeRestaurant200ResponseImage;
    street: GetMe200ResponseStreet;
    building: GetMe200ResponseStreet;
    apartment: GetMe200ResponseStreet;
    postalCode: GetMe200ResponsePostalCode;
    city: GetMe200ResponseStreet;
    menu: Array<GetMeRestaurant200ResponseMenuInner>;
}